many fixes, still not fully working
