+++ .gitignore
@@ -1,3 +1,4 @@
+/web/uploads/
 /.idea/
 /var/*
 !/var/cache
@@ -12,3 +13,4 @@
 !var/SymfonyRequirements.php
 /vendor/
 /web/bundles/
+/tests/_output/*
