+++ app/Resources/views/layout.html.twig
@@ -1,8 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
-
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -12,8 +10,8 @@
     <title>{{ page_title }}</title>
     {% block stylesheets %}
         {% stylesheets 
-            '@MemeSiteBundle/Resources/public/css/bootstrap.min.css'
-            '@MemeSiteBundle/Resources/public/css/4-col-portfolio.css'
+            '@SiteBundle/Resources/public/css/bootstrap.min.css'
+            '@SiteBundle/Resources/public/css/4-col-portfolio.css'
         %}
             <link href="{{ asset_url }}" rel="stylesheet">
         {% endstylesheets %}
@@ -40,7 +38,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a class="navbar-brand" href="{{ path('homepage') }}">Meme.tk</a>
+                <a class="navbar-brand" href="{{ path('homepage') }}">memy.tk</a>
             </div>
             <!-- Collect the nav links, forms, and other content for toggling -->
             <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
@@ -88,8 +86,8 @@
     <script src="https://code.jquery.com/jquery-1.11.3.min.js" type="javascript"></script>
     {% block javascripts %}
         {% javascripts 
-            '@MemeSiteBundle/Resources/public/js/jquery.js'
-            '@MemeSiteBundle/Resources/public/js/bootstrap.min.js'
+            '@SiteBundle/Resources/public/js/jquery.js'
+            '@SiteBundle/Resources/public/js/bootstrap.min.js'
         %}
             <script src="{{ asset_url }}"></script>
         {% endjavascripts %}
