+++ app/config/config_test.yml
@@ -1,6 +1,6 @@
 imports:
     - { resource: config_dev.yml }
-    - { resource: parameters/test.yml }
+    - { resource: parameters/loc.yml }
 
 framework:
     test: ~
