+++ composer
@@ -1,0 +1,1 @@
+bin/composer.phar
