+++ src/Meme/MemeBundle/Form/MemeType.php
@@ -2,9 +2,13 @@
 
 namespace Meme\MemeBundle\Form;
 
+use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Vich\UploaderBundle\Form\Type\VichImageType;
 
 class MemeType extends AbstractType
 {
@@ -15,18 +19,24 @@ class MemeType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('filename', 'file', [
+            ->add('image', VichImageType::class, [
                 //'multiple' => true,
-                'required' => true,
+                'required' => false,
+                'label' => 'Plik',
             ])
-            ->add('tags', 'entity', [
-                'class' => 'Meme\MemeBundle\Entity\Meme',
+            ->add('tags', EntityType::class, [
+                'class' => 'Meme\MemeBundle\Entity\Tag',
                 'multiple' => true,
                 'required' => false,
+                'label' => 'Tagi',
             ])
-            ->add('title', 'text', [
+            ->add('title', TextType::class, [
                 'required' => false,
+                'label' => 'Nazwa',
             ])
+            ->add('submit', SubmitType::class, [
+                'label' => 'Dodaj',
+            ]);
         ;
     }
     
@@ -38,13 +48,5 @@ class MemeType extends AbstractType
         $resolver->setDefaults(array(
             'data_class' => 'Meme\MemeBundle\Entity\Meme'
         ));
-    }
-
-    /**
-     * @return string
-     */
-    public function getName()
-    {
-        return 'Meme_memebundle_meme';
     }
 }
