+++ src/Meme/MemeBundle/Resources/config/routing.yml
@@ -1,3 +1,3 @@
 meme:
-    resource: "@MemeMemeBundle/Resources/config/routing/meme.yml"
+    resource: "routing/meme.yml"
     prefix:   /
