+++ src/Meme/MemeBundle/Resources/config/vich_uploader/Meme.yml
@@ -1,0 +1,4 @@
+Meme\MemeBundle\Entity\Meme:
+    image:
+        mapping:           meme
+        filename_property: filename
