+++ src/Meme/MemeBundle/Resources/views/Meme/edit.html.twig
@@ -1,4 +1,4 @@
-{% extends '::base.html.twig' %}
+{% extends '::layout.html.twig' %}
 
 {% block body -%}
     <h1>Meme edit</h1>
