+++ src/Meme/MemeBundle/Resources/views/Meme/list.html.twig
@@ -1,4 +1,4 @@
-{% extends 'MemeCoreBundle::layout.html.twig' %}
+{% extends '::layout.html.twig' %}
 
 {% block content %}
