+++ tests/_data/dump.sql
@@ -1,0 +1,1 @@
+/* Replace this file with actual dump of your database */
