+++ tests/_support/Helper/Unit.php
@@ -1,5 +1,6 @@
 <?php
 namespace Helper;
+
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
