+++ tests/_support/_generated/UnitTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] c871459f363e639d6d4c21f535bb78f4
+<?php  //[STAMP] 3bcba9e954b12f3dc227aad523b134a9
 namespace _generated;
 
 // This class was automatically generated by build task
