+++ tests/functional.suite.yml
@@ -8,6 +8,6 @@ class_name: FunctionalTester
 modules:
     enabled:
         - PhpBrowser:
-            url: http://memy.local/app_test.php
+            url: http://127.0.0.1:8000/app_test.php
         # add framework module here
         - \Helper\Functional
