+++ tests/functional/HomepageCept.php
@@ -1,6 +1,6 @@
 <?php 
-$I = new AcceptanceTester($scenario);
+$I = new FunctionalTester($scenario);
 $I->wantTo('see homepage');
 $I->amOnPage('/');
 $I->see("memy.tk");
-$I->see("twój osobisty folder z obrazkami reakcyjnymi");
+$I->see("twój osobisty folder ze śmiesznymi obrazkami");
